Do not use extra module for pkg/eraser
Multi-module repos are highly problematic for go currently and
complicate development.
Tools like gopls do not support it breaks the normal dev workflow.

Instead just keep everything in one module for now.
We may return to multi-module once the repo is stabilized or go is in a
better place for multi-module development.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>