Remove imagestatus
This replaces imagestatus with a smaller summary of the last run.
Also instead of deleting jobs immediately after completion there are
options to set a delay period that the controller should wait before
deleting.
Detailed information about what was deleted or not is output the the
eraser pod's logs now.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>