Namespace tests for eraser test client
These tests are testing the `testClient` and not actual behavior of
eraser.
This is moving them into subtests so the tests are namespaced.

There are some minor implementation changes (and one dropped test) as
the implementation was just wrong.
It was skipping "running" images in the implementation, which is not
something that the client should be doing (a real client certainly
could, but we need to be able to test our own logic).

Signed-off-by: Brian Goff <cpuguy83@gmail.com>