Skip and log non-linux nodes
If the cluster is running Windows nodes, we want to prevent scheduling
the eraser pod on any Windows nodes. In order to maintain visibility of
skipped nodes, the ImageJob will report the number of skipped nodes in
its status field. In addition, the controller manager pod will log the
event.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>