Use fixed go version to install setup-envtest
`go get` and `go install` have different behaviors and different
interpretations of environment variables such as GOBIN. Since we are
already using a go image with a fixed version to fetch `controller-gen`,
we can use the same container to fetch the `setup-envtest` binary.

This avoid the problem of `go get` behaving the wrong way if you have
the "wrong" version of go installed.

Since the tests are run locally (in a non-container environment), mount
the $PROJECT_DIR/bin directory to obtain the `setup-envtest` binary for
the local environment.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>