Generate an alternative eraser manifest
In order to register nodes that can be skipped by label, we need to
alter the kind configuration.

In addtion, we need to generate a separate manifest for testing, to
enable the command-line flag which allows us to specify the selector
that should be used in skipping.

These changes are being committed in case the code is needed later, but
I have found an alternative solution to testing which works better.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>