Stabilize e2e tests
Resolves #134

There is a lag between when a) the kubernetes client recognizes a
resource (representing a container or containers) has been deleted, and
b) containerd recognizes the containers as stopped or non-running.

Instead of polling the kubernetes api server, poll containerd directly
using `ctr` on the kind nodes. Once the container is stopped, we may
proceed with the eraser deployment.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>