Add node selector to image job pod spec (#141)
* Skip and log non-linux nodes

If the cluster is running Windows nodes, we want to prevent scheduling
the eraser pod on any Windows nodes. In order to maintain visibility of
skipped nodes, the ImageJob will report the number of skipped nodes in
its status field. In addition, the controller manager pod will log the
event.

Also adds a none label that can be applied to nodes to skip them.
Skipped jobs are considered successful as it relates to the success threshold for cleaning up image jobs.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>