Disable caching in github testing workflow
Because of rate-limiting behavior in github (see [moby/buildkit #2804](https://github.com/moby/buildkit/issues/2804)),
the caching was resuling in an exponential backoff from buildkit. This
caused the whole action to hang and subsequently time out. The e2e test
would never run, because the failure was in the build stage. This change
removes the caching entirely, which results in fewer timeout failures.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>