feat: parallelize tests (#271)
* Split e2e tests so they can be run in parallel

Move each e2e test to its own package so it can be run in parallel. The
Makefile had to be modified to accomodate the changes, as did the github
workflow for testing.

Signed-off-by: Peter Engelbert

* Disable caching in github testing workflow

Because of rate-limiting behavior in github (see [moby/buildkit #2804](https://github.com/moby/buildkit/issues/2804)),
the caching was resuling in an exponential backoff from buildkit. This
caused the whole action to hang and subsequently time out. The e2e test
would never run, because the failure was in the build stage. This change
removes the caching entirely, which results in fewer timeout failures.

Signed-off-by: Peter Engelbert

* Remove CodeQL action in github testing workflow

We now have a cron job to periodically run the CodeQL action, so it is
no longer necessary in the testing workflow. It was taking a long time,
so it has been removed here.

Signed-off-by: Peter Engelbert

* Refactor

Signed-off-by: Peter Engelbert

* Don't load scanner image

Signed-off-by: Peter Engelbert

* Load multiple kind images

Signed-off-by: Peter Engelbert

* Give tests more descriptive names and documentation

Signed-off-by: Peter Engelbert

* Split out exclusion list test after rebase

Signed-off-by: Peter Engelbert