Remove e2e test teardowns
Each test has its own teardown section, which is no longer necessary.
They used to be used in between tests, but because we are running the
tests in parallel in the CI, we are now destroying the entire cluster
after each test. We can safely remove the teardown sections from each of
the e2e tests.

Resolves #302

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>