Allow specifying namespace when getting logs in E2E test
The tests deployed by manifest use the `eraser-system` namespecace,
whereas the ones deployed by helm use a randomized test namespace.
Because of the way helm charts are generated, it's best to keep the
manifest test using `eraser-system`. This change allows that to happen
without breaking the rest of the tests.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>