test: prevent pulling default images in e2e test (#458)
Prevent pulling default images in e2e test

The file `test/e2e/test-data/helm-empty-values.yaml` exists to prevent
regression in the tests. A situation arose in which the helm keys were
wrong in the e2e tests, resulting in helm plugging in the default values
for image repository and tag. This resulted in false positives because
the cluster used in the e2e test would pull in the default image from a
registry.

For all e2e tests using helm, this file should be provided on the
command line using `helm install -f`. This ensures that the default
values are never used. The correct values will be supplied by `--set`
flags further to the right on the command line.

In that file, randomized names are used to guarantee that if the wrong
helm keys are used, the test will fail.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>