fix: label child imagejobs by owner (#457)
* Label child imagejobs by owner

Prior to PR #339, the ImageCollector controller was able to filter out
ImageJob events related to ImageJobs created by the ImageList
controller. Because the ImageCollector CR was removed, we could no
longer keep the `&handler.EnqueueRequestForOwner` condition below:

-               &handler.EnqueueRequestForOwner{OwnerType: &eraserv1alpha1.ImageCollector{}, IsController: true}, predicate.Funcs{
+               &handler.EnqueueRequestForObject{}, predicate.Funcs{

The removed line can't be restored because
`eraserv1alpha1.ImageCollector` was removed as a type in PR #339.
Therefore, we need to manually manage ownership of these imagejobs. To
accomplish this, we use an imagejob's metadata to label it. We label it
by the owning controller, and then check the label as a precondition for
reconciling on an ImageJob event.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Fix errors

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Update e2e test to ensure eraser imagejobs aren't recreated

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Remove commented code

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>