Add default tolerations for collector/eraser pods
Issue #503 points out that pods will not be scheduled on nodes with
taints, because we do not set any tolerations on the collector or eraser
pods.

By contrast, daemonsets have default tolerations that enable them to be
scheduled on every node regardless of taints. This PR applies those
defaults to the collector and eraser pods.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>