Do not remove failed images in e2e tests
This commit introduces an easier-to-use way of setting additional args
on the manager for the e2e tests. Along with this change,
`--scanner-arg=--delete-scan-failed-images=false` is the default,
meaning that tests will not delete images for which the scan failed.

This is to address #536, where a failing scanner was not caught by the
e2e tests.

I have verified locally that a failing scanner will cause at least one
test to fail.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>