test: catch broken scanner (#540)
* Do not remove failed images in e2e tests

This commit introduces an easier-to-use way of setting additional args
on the manager for the e2e tests. Along with this change,
`--scanner-arg=--delete-scan-failed-images=false` is the default,
meaning that tests will not delete images for which the scan failed.

This is to address #536, where a failing scanner was not caught by the
e2e tests.

I have verified locally that a failing scanner will cause at least one
test to fail.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Ensure failing scanner causes at least one e2e test failure

Adds a new e2e test

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Rename tests to sort logs

Multiple tests were assigned the same name, and because the test_logs
are placed in directories named after the tests, the log directories
were getting cluttered with results from multiple tests. This commit
gives each test a unique name, and therefore a unique log directory.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Tidy code in imagecollector_controller

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>