chore: Use PodTemplates to support ImageJobs (#555)
* Use PodTemplates to support ImageJobs

This PR was caused by limits on the size of k8s API requests. By
default, `kubectl apply` will add a field, `metadata.annotations.kubectl."kubernetes.io/last-applied-configuration"`,
which is a JSON representation of the object. This effectively increases
the size of the request.

We ran into size limits when adding a `v1` of the eraser API. Because we
want to support a multi-version API, the request to create the CRD for
imagejobs was essentially doubled by the addition of the new version.

The component of the CRD that took up the most space was the
`ImageJob.Spec` field, a `PodTemplateSpec` type. The generated YAML for
this type is extremely long, with a lot of documentation. Being a core
kubernetes resource, a `PodTemplateSpec` is very long and complex.

This commit leverages the kubernetes resource `PodTemplate` in order to
remove the `ImageJob.Spec` field entirely. Instead of using the
`ImageJob` to specify configuration of the `eraser` or `collector` pods,
that responsibility is offloaded to a `PodTemplate`.

There is a 1:1 ratio between imagejobs and podtemplates. A given
imagejob is expected to have a corresponding podtemplate with the same
name, in the same namespace as the controller-manager pod. If no such
podtemplate exists, an error is returned and the reconciliation loop
spins until the podtemplate is found.

The OwnerReference of the podtemplate is set to its corresponding
imagejob. When the imagejob is deleted, the podtemplate will be cleaned
up as well.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Address failure in imagelist_rm_images e2e test

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Test without automated collector

This code should be reverted once the timing logic has been sorted out.
See issue #568 for details.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Fix error in initial trigger logic

Also, remove unnecessary goroutine.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Address PR comments

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Address more PR comments

- rename `repeatImmediate` to `scheduleImmediate`
- remove kubebuilder comment for k8s `jobs` which we no longer use

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Re-generate manifests

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

* Fix typo in e2e test

Mistakenly had the old flag, `--repeat-immediate` instead of the new,
`--schedule-immediate`

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>