Restore ability to catch broken scanner
Prior to the configmap changes, there was an e2e test,
`collector_ensure_scan` designed to catch when the scanner is broken.
The reason this is needed is that when the scanner isn't functioning
properly, and `deleteFailedImages` is set to `true`, then scanning of
all images will fail.

Because in most tests we delete those images for which scanning failed,
this will create a false positive: *all* images will be removed, instead
of just the vulnerable ones. That makes it impossible to tell whether or
not the tests should be passing, because deleting all images will remove
the vulnerable images along with the safe ones.

This PR restores the test to a working state, and is critical to catch
when we break the scanner.

I have manually tested and confirmed that the modified test fails when
the scanner is broken.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>