Fix build/version.sh to accept BuildVersion
The build/version.sh script was setting `buildVersion` instead of
`BuildVersion`. This was causing the default values for tags to be empty
strings, which is what we want, but not how we want to get there.

The tags will be provided by either the helm chart or the manifest.
Thus, in order for them to be overwritten during unmarshaling, they must
be empty in the default config.

This commit explicitly makes the image tags empty.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>