Install trivy binary into trivy-scanner image
Installs the published trivy binary to the trivy-scanner image. Make
variables TRIVY_BINARY_REPO TRIVY_BINARY_TAG, and TRIVY_BINARY_IMG have
been added to provide control over the versioning.

This is a beginning of a change intended to reduce maintentance and
to resolve resource issues that have arisen. Instead of using trivy as a
library (which has a constantly-shifting API), we will use the published
binary.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>