Revise structure of socket config
Because trivy's behavior switches on the name of the runtime, it has to
be supplied at the time trivy runs.

To specify the address, there are 3 different environment variables that
trivy will check in order to find the socket, depending on the runtime.

In the case of the crio runtime, an exact address cannot be specified
(or rather, trivy won't understand it). Trivy will always look for the
podman address at $XDG_RUNTIME_DIR/podman/podman.sock . While the value
of XDG_RUNTIME_DIR can be changed, the exact socket address can't be.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>